consumer/requirements.txt
aiofiles==0.4.0
aiokafka==0.4.3
aiopg==0.15.0
cassandra-driver==3.16.0
httptools==0.0.11
kafka-python==1.4.3
kazoo==2.6.0
marshmallow==2.16.3
psycopg2==2.7.6.1
redis==3.0.1
sanic==0.7.0
six==1.12.0
SQLAlchemy==1.2.15
ujson==1.35
uvloop==0.11.3
websockets==7.0
